fix: stabilize interrupted-recording and remote e2e tests
Preserve active recordings when rebinding the mixer harness during resume
validation, assert WAV growth instead of stale session.json metadata, and
reorder remote client playback before recording. Fix JSONObject.getFloat
crash on Android when decoding remote waveform chunks.

+++ remote-server/src/main/kotlin/org/openmultitrack/remote/RemoteJson.kt
@@ -119,7 +119,7 @@ internal object RemoteJson {
             mixerId = root.optString("mixerId", ""),
             sessionDir = root.optString("sessionDir", ""),
             channel = root.getInt("channel"),
-            startSec = root.getFloat("startSec"),
+            startSec = root.getDouble("startSec").toFloat(),
             peaks = peaks,
         )
     }
